requirements.txt
groq
streamlit
python-dotenv
openpyxl
python-docx
pdfplumber